Add tests for check command

+++ tests/temp_dir.py
@@ -5,9 +5,10 @@ from pathlib import Path
 
 
 @contextmanager
-def TEMP_DIR(root: Path = Path(".")):
+def TEMP_DIR(root: Path = Path("."), delete: bool = True):
     temp_dir = root / ".temp"
     shutil.rmtree(temp_dir, True)
     temp_dir.mkdir(0o777, True, True)
     yield temp_dir
-    shutil.rmtree(temp_dir, True)
+    if delete:
+        shutil.rmtree(temp_dir, True)
